requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==4.0.10
django-allauth==0.48.0
django-cors-headers==3.10.1
django-extensions==3.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
drf-spectacular==0.21.2
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
psycopg2-binary
python-dotenv